Main 0.3v

+++ packages/server/src/controllers/dataset/index.ts
@@ -7,8 +7,14 @@ import { getPageAndLimitParams } from '../../utils/pagination'
 const getAllDatasets = async (req: Request, res: Response, next: NextFunction) => {
     try {
         const { page, limit } = getPageAndLimitParams(req)
-        const workspaceId = req.user?.activeWorkspaceId || ''
-const apiResponse = await datasetService.getAllDatasets(workspaceId, page, limit)
+        const workspaceId = req.user?.activeWorkspaceId
+        if (!workspaceId) {
+            throw new InternalFlowiseError(
+                StatusCodes.NOT_FOUND,
+                `Error: datasetController.getAllDatasets - workspace ${workspaceId} not found!`
+            )
+        }
+        const apiResponse = await datasetService.getAllDatasets(workspaceId, page, limit)
         return res.json(apiResponse)
     } catch (error) {
         next(error)
@@ -21,8 +27,14 @@ const getDataset = async (req: Request, res: Response, next: NextFunction) => {
             throw new InternalFlowiseError(StatusCodes.PRECONDITION_FAILED, `Error: datasetService.getDataset - id not provided!`)
         }
         const { page, limit } = getPageAndLimitParams(req)
-        const workspaceId = req.user?.activeWorkspaceId || ''
-const apiResponse = await datasetService.getDataset(req.params.id, workspaceId, page, limit)
+        const workspaceId = req.user?.activeWorkspaceId
+        if (!workspaceId) {
+            throw new InternalFlowiseError(
+                StatusCodes.NOT_FOUND,
+                `Error: datasetController.getDataset - workspace ${workspaceId} not found!`
+            )
+        }
+        const apiResponse = await datasetService.getDataset(req.params.id, workspaceId, page, limit)
         return res.json(apiResponse)
     } catch (error) {
         next(error)
@@ -35,8 +47,14 @@ const createDataset = async (req: Request, res: Response, next: NextFunction) =>
             throw new InternalFlowiseError(StatusCodes.PRECONDITION_FAILED, `Error: datasetService.createDataset - body not provided!`)
         }
         const body = req.body
-        const workspaceId = req.user?.activeWorkspaceId || ''
-body.workspaceId = workspaceId
+        const workspaceId = req.user?.activeWorkspaceId
+        if (!workspaceId) {
+            throw new InternalFlowiseError(
+                StatusCodes.NOT_FOUND,
+                `Error: datasetController.createDataset - workspace ${workspaceId} not found!`
+            )
+        }
+        body.workspaceId = workspaceId
         const apiResponse = await datasetService.createDataset(body)
         return res.json(apiResponse)
     } catch (error) {
@@ -52,8 +70,14 @@ const updateDataset = async (req: Request, res: Response, next: NextFunction) =>
         if (typeof req.params === 'undefined' || !req.params.id) {
             throw new InternalFlowiseError(StatusCodes.PRECONDITION_FAILED, `Error: datasetService.updateDataset - id not provided!`)
         }
-        const workspaceId = req.user?.activeWorkspaceId || ''
-const apiResponse = await datasetService.updateDataset(req.params.id, req.body, workspaceId)
+        const workspaceId = req.user?.activeWorkspaceId
+        if (!workspaceId) {
+            throw new InternalFlowiseError(
+                StatusCodes.NOT_FOUND,
+                `Error: datasetController.updateDataset - workspace ${workspaceId} not found!`
+            )
+        }
+        const apiResponse = await datasetService.updateDataset(req.params.id, req.body, workspaceId)
         return res.json(apiResponse)
     } catch (error) {
         next(error)
@@ -65,8 +89,14 @@ const deleteDataset = async (req: Request, res: Response, next: NextFunction) =>
         if (typeof req.params === 'undefined' || !req.params.id) {
             throw new InternalFlowiseError(StatusCodes.PRECONDITION_FAILED, `Error: datasetService.deleteDataset - id not provided!`)
         }
-        const workspaceId = req.user?.activeWorkspaceId || ''
-const apiResponse = await datasetService.deleteDataset(req.params.id, workspaceId)
+        const workspaceId = req.user?.activeWorkspaceId
+        if (!workspaceId) {
+            throw new InternalFlowiseError(
+                StatusCodes.NOT_FOUND,
+                `Error: datasetController.deleteDataset - workspace ${workspaceId} not found!`
+            )
+        }
+        const apiResponse = await datasetService.deleteDataset(req.params.id, workspaceId)
         return res.json(apiResponse)
     } catch (error) {
         next(error)
@@ -81,8 +111,14 @@ const addDatasetRow = async (req: Request, res: Response, next: NextFunction) =>
         if (!req.body.datasetId) {
             throw new InternalFlowiseError(StatusCodes.PRECONDITION_FAILED, `Error: datasetService.addDatasetRow - datasetId not provided!`)
         }
-        const workspaceId = req.user?.activeWorkspaceId || ''
-req.body.workspaceId = workspaceId
+        const workspaceId = req.user?.activeWorkspaceId
+        if (!workspaceId) {
+            throw new InternalFlowiseError(
+                StatusCodes.NOT_FOUND,
+                `Error: datasetController.addDatasetRow - workspace ${workspaceId} not found!`
+            )
+        }
+        req.body.workspaceId = workspaceId
         const apiResponse = await datasetService.addDatasetRow(req.body)
         return res.json(apiResponse)
     } catch (error) {
@@ -98,8 +134,14 @@ const updateDatasetRow = async (req: Request, res: Response, next: NextFunction)
         if (typeof req.params === 'undefined' || !req.params.id) {
             throw new InternalFlowiseError(StatusCodes.PRECONDITION_FAILED, `Error: datasetService.updateDatasetRow - id not provided!`)
         }
-        const workspaceId = req.user?.activeWorkspaceId || ''
-req.body.workspaceId = workspaceId
+        const workspaceId = req.user?.activeWorkspaceId
+        if (!workspaceId) {
+            throw new InternalFlowiseError(
+                StatusCodes.NOT_FOUND,
+                `Error: datasetController.updateDatasetRow - workspace ${workspaceId} not found!`
+            )
+        }
+        req.body.workspaceId = workspaceId
         const apiResponse = await datasetService.updateDatasetRow(req.params.id, req.body)
         return res.json(apiResponse)
     } catch (error) {
@@ -112,8 +154,14 @@ const deleteDatasetRow = async (req: Request, res: Response, next: NextFunction)
         if (typeof req.params === 'undefined' || !req.params.id) {
             throw new InternalFlowiseError(StatusCodes.PRECONDITION_FAILED, `Error: datasetService.deleteDatasetRow - id not provided!`)
         }
-        const workspaceId = req.user?.activeWorkspaceId || ''
-const apiResponse = await datasetService.deleteDatasetRow(req.params.id, workspaceId)
+        const workspaceId = req.user?.activeWorkspaceId
+        if (!workspaceId) {
+            throw new InternalFlowiseError(
+                StatusCodes.NOT_FOUND,
+                `Error: datasetController.deleteDatasetRow - workspace ${workspaceId} not found!`
+            )
+        }
+        const apiResponse = await datasetService.deleteDatasetRow(req.params.id, workspaceId)
         return res.json(apiResponse)
     } catch (error) {
         next(error)
@@ -123,8 +171,14 @@ const apiResponse = await datasetService.deleteDatasetRow(req.params.id, workspa
 const patchDeleteRows = async (req: Request, res: Response, next: NextFunction) => {
     try {
         const ids = req.body.ids ?? []
-        const workspaceId = req.user?.activeWorkspaceId || ''
-const apiResponse = await datasetService.patchDeleteRows(ids, workspaceId)
+        const workspaceId = req.user?.activeWorkspaceId
+        if (!workspaceId) {
+            throw new InternalFlowiseError(
+                StatusCodes.NOT_FOUND,
+                `Error: datasetController.patchDeleteRows - workspace ${workspaceId} not found!`
+            )
+        }
+        const apiResponse = await datasetService.patchDeleteRows(ids, workspaceId)
         return res.json(apiResponse)
     } catch (error) {
         next(error)
@@ -136,8 +190,14 @@ const reorderDatasetRow = async (req: Request, res: Response, next: NextFunction
         if (!req.body) {
             throw new InternalFlowiseError(StatusCodes.PRECONDITION_FAILED, `Error: datasetService.reorderDatasetRow - body not provided!`)
         }
-        const workspaceId = req.user?.activeWorkspaceId || ''
-const apiResponse = await datasetService.reorderDatasetRow(req.body.datasetId, req.body.rows, workspaceId)
+        const workspaceId = req.user?.activeWorkspaceId
+        if (!workspaceId) {
+            throw new InternalFlowiseError(
+                StatusCodes.NOT_FOUND,
+                `Error: datasetController.reorderDatasetRow - workspace ${workspaceId} not found!`
+            )
+        }
+        const apiResponse = await datasetService.reorderDatasetRow(req.body.datasetId, req.body.rows, workspaceId)
         return res.json(apiResponse)
     } catch (error) {
         next(error)
